refactor: reorganize test structure and remove Weaviate references
- Move test utilities from src/Testing/ to tests/Support/
- Reorganize integration tests to follow namespace structure:
  - tests/Integration/Registry/ for EdgeBinder\Registry\* tests
  - tests/Integration/Support/ for test factories
- Remove all Weaviate-specific code and references from core project:
  - Delete WeaviateLikeTransformer and Weaviate integration tests
  - Update AbstractAdapterTestSuite comments to be adapter-agnostic
  - Replace Weaviate examples with generic adapter examples
- Update namespaces for moved classes to EdgeBinder\Tests\*
- Keep AbstractAdapterTestSuite in src/Testing/ (used by external adapters)

This cleanup ensures the core EdgeBinder project contains no adapter-specific
code while maintaining proper test organization following namespace conventions.
