+++ tests/Integration/Support/InMemoryAdapterFactory.php
@@ -2,33 +2,28 @@
 
 declare(strict_types=1);
 
-namespace EdgeBinder\Tests\Integration;
+namespace EdgeBinder\Tests\Integration\Support;
 
 use EdgeBinder\Contracts\PersistenceAdapterInterface;
 use EdgeBinder\Persistence\InMemory\InMemoryAdapter;
 use EdgeBinder\Registry\AdapterFactoryInterface;
 
 /**
- * Mock adapter factory for integration testing.
+ * InMemory adapter factory for integration testing.
  *
  * This factory creates InMemory adapters that can be used to test the
  * adapter registry integration with real storage functionality.
  */
-final class MockAdapterFactory implements AdapterFactoryInterface
+final class InMemoryAdapterFactory implements AdapterFactoryInterface
 {
     public function __construct(
-        private readonly string $adapterType = 'mock',
-        private readonly ?PersistenceAdapterInterface $adapter = null
+        private readonly string $adapterType = 'inmemory'
     ) {
     }
 
     public function createAdapter(array $config): PersistenceAdapterInterface
     {
-        // Return pre-configured adapter if provided, otherwise create a new InMemory adapter
-        if (null !== $this->adapter) {
-            return $this->adapter;
-        }
-
+        // The InMemory adapter doesn't need any configuration
         return new InMemoryAdapter();
     }
