+++ tests/Integration/Support/MockAdapterFactory.php
@@ -2,28 +2,33 @@
 
 declare(strict_types=1);
 
-namespace EdgeBinder\Tests\Integration;
+namespace EdgeBinder\Tests\Integration\Support;
 
 use EdgeBinder\Contracts\PersistenceAdapterInterface;
 use EdgeBinder\Persistence\InMemory\InMemoryAdapter;
 use EdgeBinder\Registry\AdapterFactoryInterface;
 
 /**
- * InMemory adapter factory for integration testing.
+ * Mock adapter factory for integration testing.
  *
  * This factory creates InMemory adapters that can be used to test the
  * adapter registry integration with real storage functionality.
  */
-final class InMemoryAdapterFactory implements AdapterFactoryInterface
+final class MockAdapterFactory implements AdapterFactoryInterface
 {
     public function __construct(
-        private readonly string $adapterType = 'inmemory'
+        private readonly string $adapterType = 'mock',
+        private readonly ?PersistenceAdapterInterface $adapter = null
     ) {
     }
 
     public function createAdapter(array $config): PersistenceAdapterInterface
     {
-        // The InMemory adapter doesn't need any configuration
+        // Return pre-configured adapter if provided, otherwise create a new InMemory adapter
+        if (null !== $this->adapter) {
+            return $this->adapter;
+        }
+
         return new InMemoryAdapter();
     }
