+++ tests/Support/InMemoryEdgeBinder.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EdgeBinder\Testing;
+namespace EdgeBinder\Tests\Support;
 
 use EdgeBinder\Binding;
 use EdgeBinder\Contracts\BindingInterface;
